+++ CachingHttpClient.php
@@ -18,10 +18,6 @@ use Symfony\Component\HttpClient\Response\AsyncContext;
 use Symfony\Component\HttpClient\Response\AsyncResponse;
 use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Component\HttpClient\Response\ResponseStream;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\HttpCache\HttpCache;
-use Symfony\Component\HttpKernel\HttpCache\StoreInterface;
-use Symfony\Component\HttpKernel\HttpClientKernel;
 use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\Cache\TagAwareCacheInterface;
 use Symfony\Contracts\HttpClient\ChunkInterface;
@@ -96,7 +92,6 @@ class CachingHttpClient implements HttpClientInterface, ResetInterface
      */
     private const MAX_HEURISTIC_FRESHNESS_TTL = 86400;
 
-    private TagAwareCacheInterface|HttpCache $cache;
     private array $defaultOptions = self::OPTIONS_DEFAULTS;
 
     /**
@@ -109,35 +104,11 @@ class CachingHttpClient implements HttpClientInterface, ResetInterface
      */
     public function __construct(
         private HttpClientInterface $client,
-        TagAwareCacheInterface|StoreInterface $cache,
+        private readonly TagAwareCacheInterface $cache,
         array $defaultOptions = [],
         private readonly bool $sharedCache = true,
         private readonly ?int $maxTtl = null,
     ) {
-        if ($cache instanceof StoreInterface) {
-            trigger_deprecation('symfony/http-client', '7.4', 'Passing a "%s" as constructor\'s 2nd argument of "%s" is deprecated, "%s" expected.', StoreInterface::class, __CLASS__, TagAwareCacheInterface::class);
-
-            if (!class_exists(HttpClientKernel::class)) {
-                throw new \LogicException(\sprintf('Using "%s" requires the HttpKernel component, try running "composer require symfony/http-kernel".', __CLASS__));
-            }
-
-            $kernel = new HttpClientKernel($client);
-            $this->cache = new HttpCache($kernel, $cache, null, $defaultOptions);
-
-            unset($defaultOptions['debug']);
-            unset($defaultOptions['default_ttl']);
-            unset($defaultOptions['private_headers']);
-            unset($defaultOptions['skip_response_headers']);
-            unset($defaultOptions['allow_reload']);
-            unset($defaultOptions['allow_revalidate']);
-            unset($defaultOptions['stale_while_revalidate']);
-            unset($defaultOptions['stale_if_error']);
-            unset($defaultOptions['trace_level']);
-            unset($defaultOptions['trace_header']);
-        } else {
-            $this->cache = $cache;
-        }
-
         if ($defaultOptions) {
             [, $this->defaultOptions] = self::prepareRequest(null, null, $defaultOptions, $this->defaultOptions);
         }
@@ -145,10 +116,6 @@ class CachingHttpClient implements HttpClientInterface, ResetInterface
 
     public function request(string $method, string $url, array $options = []): ResponseInterface
     {
-        if ($this->cache instanceof HttpCache) {
-            return $this->legacyRequest($method, $url, $options);
-        }
-
         [$fullUrl, $options] = self::prepareRequest($method, $url, $options, $this->defaultOptions);
 
         $fullUrl = implode('', $fullUrl);
@@ -422,46 +389,6 @@ class CachingHttpClient implements HttpClientInterface, ResetInterface
             yield from MockResponse::stream($mockResponses, $timeout);
             yield from $this->asyncStream($asyncResponses, $timeout);
         })());
-    }
-
-    private function legacyRequest(string $method, string $url, array $options = []): ResponseInterface
-    {
-        [$url, $options] = self::prepareRequest($method, $url, $options, $this->defaultOptions, true);
-        $url = implode('', $url);
-
-        if (!empty($options['body']) || !empty($options['extra']['no_cache']) || !\in_array($method, ['GET', 'HEAD', 'OPTIONS'], true)) {
-            return new AsyncResponse($this->client, $method, $url, $options);
-        }
-
-        $request = Request::create($url, $method);
-        $request->attributes->set('http_client_options', $options);
-
-        foreach ($options['normalized_headers'] as $name => $values) {
-            if ('cookie' !== $name) {
-                foreach ($values as $value) {
-                    $request->headers->set($name, substr($value, 2 + \strlen($name)), false);
-                }
-
-                continue;
-            }
-
-            foreach ($values as $cookies) {
-                foreach (explode('; ', substr($cookies, \strlen('Cookie: '))) as $cookie) {
-                    if ('' !== $cookie) {
-                        $cookie = explode('=', $cookie, 2);
-                        $request->cookies->set($cookie[0], $cookie[1] ?? '');
-                    }
-                }
-            }
-        }
-
-        $response = $this->cache->handle($request);
-        $response = new MockResponse($response->getContent(), [
-            'http_code' => $response->getStatusCode(),
-            'response_headers' => $response->headers->allPreserveCase(),
-        ]);
-
-        return MockResponse::fromRequest($method, $url, $options, $response);
     }
 
     private static function hash(string $toHash): string